+++ example/lib/main.dart
@@ -1,6 +1,4 @@
 // ignore_for_file: avoid_redundant_argument_values
-import 'dart:developer' show log;
-
 import 'package:dio/dio.dart';
 import 'package:flutter/foundation.dart' show kDebugMode;
 import 'package:omega_dio_logger/omega_dio_logger.dart';
@@ -10,17 +8,18 @@ void main() {
 
   dio.interceptors.add(
     const OmegaDioLogger(
-      logPrint: log,
+      error: true,
+      request: true,
+      requestHeader: true,
+      requestQueryParameters: true,
+      requestBody: true,
+      response: true,
+      responseHeader: true,
+      responseBody: true,
       convertFormData: true,
-      showError: true,
-      showRequest: true,
-      showRequestBody: true,
-      showRequestHeaders: true,
-      showRequestQueryParameters: true,
-      showResponse: true,
-      showResponseBody: true,
-      showResponseHeaders: true,
+      colorized: true,
       showCurl: true,
+      logPrint: print,
       showLog: kDebugMode,
     ),
   );