+++ CHANGELOG.md
@@ -1,10 +1,14 @@
+## 1.1.0
+
+- Add color to messages
+
 ## 1.0.2
 
-- Fixed error with converting MultipartFile
+- Fixed error with converting `MultipartFile`
 
 ## 1.0.1
 
-- Support Dart '>=2.15.0 < 4.0.0'
+- Support Dart '>=2.15.0 <4.0.0'
 
 ## 1.0.0